Move InputNumber into Opal
The stepper number input. Its wrapper chrome ports to a component
stylesheet as data-variant rules kept in step with inputs/shared.css;
the inner text classes were plain utilities and inline into one map. The
className prop had no consumers and is dropped per Opal's WithoutStyles
convention.
