+++ web/lib/opal/src/components/inputs/input-number/InputNumber.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import InputNumber from "./InputNumber";
+import InputNumber from "@opal/components/inputs/input-number/components";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 
 const meta: Meta<typeof InputNumber> = {
-  title: "refresh-components/inputs/InputNumber",
+  title: "opal/components/InputNumber",
   component: InputNumber,
   tags: ["autodocs"],
   decorators: [
